requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.7
certifi==2020.12.5
comtypes==1.1.8
distlib==0.3.1
Django==3.1.5
djangorestframework==3.12.4
filelock==3.0.12
importlib-metadata==3.7.3
MouseInfo==0.1.3
numpy==1.19.4
pandas==1.1.4
Pillow==8.1.2
pipenv==2020.11.15
PyAutoGUI==0.9.52
pycodestyle==2.7.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.4
pywin32==300
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
virtualenv==20.4.3
virtualenv-clone==0.5.4
zipp==3.4.1
